+++ Configuration GUI/src/App.tsx
@@ -33,8 +33,10 @@ import Settings from "./sites/Settings";
 import Offline from "./sites/Offline"
 import Tables from "./sites/Tables";
 import config from "./config";
-import {DefaultStream, DefaultStreamSettings, StreamSettings} from "./common/Interfaces";
+import {StreamSettings} from "./common/Interfaces";
 import {Stream} from "./common/Interfaces";
+import {validateStreams, validateStreamSettings} from "./common/Validators";
+
 
 const App = () => {
     const [error, set_error] = useState(false)
@@ -55,28 +57,25 @@ const App = () => {
 
     `
 
+
+
     // Validates the stored streams and stream settings in the local storage
     // Clears local storage if some streams/settings are not valid
     // This may be needed if the UI got an update (new stream properties), but the local storage
     // holds "old" streams/settings without the new property
     const validateLocalStorage = () => {
-        const defaultStream = DefaultStream(1)
-        const defaultStreamSetting = DefaultStreamSettings(1, 5)
-
         try {
             let stored_streams: Stream[] = JSON.parse(localStorage.getItem("streams") ?? "[]")
             let stored_settings: StreamSettings[] = JSON.parse(localStorage.getItem("streamSettings") ?? "[]")
 
-            if(!stored_streams.every(s => Object.keys(defaultStream).every(key => Object.keys(s).includes(key)))) {
+            if(!validateStreams(stored_streams)) {
                 alert("Incompatible stream description found. This may be due to an update. Resetting local storage.")
                 localStorage.clear()
                 window.location.reload()
                 return
             }
 
-            if(!stored_settings.every(s => Object.keys(defaultStreamSetting).every(key => {
-                return Object.keys(s).includes(key) && s.mpls_stack != undefined
-            }))) {
+            if(!validateStreamSettings(stored_settings)) {
                 alert("Incompatible stream description found. This may be due to an update. Resetting local storage.")
                 localStorage.clear()
                 window.location.reload()